from features - run/skip option for laplacian smoothing

+++ ui/Rhino/RV2/dev/RV2pattern_from_features_cmd.py
@@ -29,37 +29,46 @@ def RunCommand(is_interactive):
         return
 
     crv_guids = []
-    pt_guids = compas_rhino.select_points("Optional. Select points for pole singularities.") or []
+    pt_guids = compas_rhino.select_points("Step 1/3 (Optional) - Select points for pole singularities.") or []
 
     box = compas_rhino.rs.BoundingBox([srf_guid])
     input_subdivision_spacing = 0.01 * compas_rhino.rs.Distance(box[0], box[6])
 
-    mesh_edge_length = compas_rhino.rs.GetReal("Pattern edge-length target.", 1.0)
+    mesh_edge_length = compas_rhino.rs.GetReal("Step 2/3 - Enter target length for edges.", 1.0)
 
     delaunay = proxy.function("compas.geometry.triangulation.triangulation_numpy.delaunay_from_points_numpy")
 
-    # print("Decompose surface and generate a pattern...")
     pattern = Pattern.from_surface_and_features(input_subdivision_spacing, mesh_edge_length, srf_guid, crv_guids, pt_guids, delaunay)
-    # print("Pattern topology generated.")
-
-    # print("Relax pattern on surface...")
-    kmax = compas_rhino.rs.GetInteger("Number of iterations for constrained Laplacian smoothing.", 50)
-    constraints = automated_smoothing_surface_constraints(pattern, srf_guid)
-    constraints.update(automated_smoothing_constraints(pattern, points=pt_guids, curves=crv_guids))
-    constrained_smoothing(pattern, kmax=kmax, damping=0.5, constraints=constraints, algorithm='area')
-    objs = set(constraints.values())
-    inputs = [srf_guid] + crv_guids + pt_guids
-    for obj in objs:
-        if obj not in inputs:
-            compas_rhino.rs.DeleteObject(obj)
-    for guid in inputs:
-        compas_rhino.rs.HideObject(guid)
-    # print("Pattern relaxed on surface.")
-
-    compas_rhino.rs.HideObjects([srf_guid] + crv_guids + pt_guids)
 
     scene.clear()
     scene.add(pattern, name='pattern')
+    scene.update()
+
+    kmax = 10
+
+    while True:
+        option = compas_rhino.rs.GetString("Step 3/3 (Optional) - Press Enter to run constrained Laplacian smoothing or ESC to skip.", strings=['Iterations'])
+
+        if option is None:
+            break
+
+        if not option:
+            constraints = automated_smoothing_surface_constraints(pattern, srf_guid)
+            constraints.update(automated_smoothing_constraints(pattern, points=pt_guids, curves=crv_guids))
+            constrained_smoothing(pattern, kmax=kmax, damping=0.5, constraints=constraints, algorithm='area')
+            objs = set(constraints.values())
+            inputs = [srf_guid] + crv_guids + pt_guids
+            for obj in objs:
+                if obj not in inputs:
+                    compas_rhino.rs.DeleteObject(obj)
+            compas_rhino.rs.HideObjects(inputs)
+            break
+
+        if option == 'Iterations':
+            new_kmax = compas_rhino.rs.GetInteger("Number of iterations:", kmax)
+            if new_kmax or new_kmax is not None:
+                kmax = new_kmax
+
     scene.update()
 
     print('Pattern object successfully created. Input object has been hidden.')
